+++ src/lib/components/colorpicker/index.vue
@@ -48,8 +48,8 @@
                     'background-color' : colorH
                 }"
 
-                @mouseup="_hslHSync(true)"
                 @mousedown="_pickColor($event);_hslHSync(false)"
+                @mouseup="_hslHSync(true)"
             >
                 <div class="mask-white"></div>
                 <div class="mask-black"></div>
@@ -61,6 +61,7 @@
                         top : data.straw.y + 'px'
                     }"
                     @mousedown.capture="_moveStraw"
+                    @mouseup.captrue="_stopmoveStraw"
                 ></div>
             </div>
             <div class="tools">
@@ -92,6 +93,7 @@
                         <morning-slider 
                             :show-tip="false" 
                             :max="360"
+                            :step="0.01"
                             :state="inputIsReadonly ? 'readonly' : 'normal'"
 
                             v-model="data.hslHReversal"
@@ -107,8 +109,8 @@
 
                             v-model="data.alpha"
 
-                            @mouseup="_hslHSync(true)"
                             @mousedown="_hslHSync(false)"
+                            @mouseup="_hslHSync(true)"
                             @value-change="_alphaChange"
                         ></morning-slider>
                     </div>
@@ -117,10 +119,10 @@
             <div class="values">
                 <div class="input">
                     <div class="rgba" v-if="data.showValueType === 'rgba' && typeof data.colorValue === 'object'">
-                        <morning-textinput v-model="data.colorValue.r" :state="inputIsReadonly ? 'readonly' : 'normal'" @value-change="_rgbaChangeR"></morning-textinput>
-                        <morning-textinput v-model="data.colorValue.g" :state="inputIsReadonly ? 'readonly' : 'normal'" @value-change="_rgbaChangeG"></morning-textinput>
-                        <morning-textinput v-model="data.colorValue.b" :state="inputIsReadonly ? 'readonly' : 'normal'" @value-change="_rgbaChangeB"></morning-textinput>
-                        <morning-textinput v-model="data.colorValue.a" :state="(inputIsReadonly || !conf.allowAlpha) ? 'readonly' : 'normal'" @value-change="_alphaChangePer"></morning-textinput>
+                        <morning-textinput v-model="data.colorValue.r" :state="inputIsReadonly ? 'readonly' : 'normal'" @focus="_hslHSync(true)" @blur="_rgbaChangeR();_hslHSync(false)"></morning-textinput>
+                        <morning-textinput v-model="data.colorValue.g" :state="inputIsReadonly ? 'readonly' : 'normal'" @focus="_hslHSync(true)" @blur="_rgbaChangeG();_hslHSync(false)"></morning-textinput>
+                        <morning-textinput v-model="data.colorValue.b" :state="inputIsReadonly ? 'readonly' : 'normal'" @focus="_hslHSync(true)" @blur="_rgbaChangeB();_hslHSync(false)"></morning-textinput>
+                        <morning-textinput v-model="data.colorValue.a" :state="(inputIsReadonly || !conf.allowAlpha) ? 'readonly' : 'normal'" @blur="_alphaChangePer"></morning-textinput>
                         <div class="name">R</div>
                         <div class="name">G</div>
                         <div class="name">B</div>
@@ -131,14 +133,16 @@
                             v-model="data.colorValue"
                             :state="inputIsReadonly ? 'readonly' : 'normal'"
                             @value-change="_hexChange"
+                            @focus="_hslHSync(true)"
+                            @blur="_hslHSync(false)"
                         ></morning-textinput>
                         <div class="name">HEX</div>
                     </div>
                     <div class="hsla" v-if="data.showValueType === 'hsla' && typeof data.colorValue === 'object'">
-                        <morning-textinput v-model="data.colorValue.h" :state="inputIsReadonly ? 'readonly' : 'normal'" @value-change="_hslChangeH"></morning-textinput>
-                        <morning-textinput v-model="data.colorValue.s" :state="inputIsReadonly ? 'readonly' : 'normal'" @value-change="_hslChangeS"></morning-textinput>
-                        <morning-textinput v-model="data.colorValue.l" :state="inputIsReadonly ? 'readonly' : 'normal'" @value-change="_hslChangeL"></morning-textinput>
-                        <morning-textinput v-model="data.colorValue.a" :state="(inputIsReadonly || !conf.allowAlpha) ? 'readonly' : 'normal'" @value-change="_alphaChangePer"></morning-textinput>
+                        <morning-textinput v-model="data.colorValue.h" :state="inputIsReadonly ? 'readonly' : 'normal'" @focus="_hslHSync(true)" @blur="_hslChangeH();_hslHSync(false)"></morning-textinput>
+                        <morning-textinput v-model="data.colorValue.s" :state="inputIsReadonly ? 'readonly' : 'normal'" @blur="_hslChangeS"></morning-textinput>
+                        <morning-textinput v-model="data.colorValue.l" :state="inputIsReadonly ? 'readonly' : 'normal'" @blur="_hslChangeL"></morning-textinput>
+                        <morning-textinput v-model="data.colorValue.a" :state="(inputIsReadonly || !conf.allowAlpha) ? 'readonly' : 'normal'" @blur="_alphaChangePer"></morning-textinput>
                         <div class="name">H</div>
                         <div class="name">S</div>
                         <div class="name">L</div>
@@ -174,6 +178,7 @@ const num16 = 16;
 const num100 = 100;
 const num360 = 360;
 const maxAlpha = 255;
+const maxHex = 255;
 const defaultColor = '#000000';
 const valueTypes = [
     'hex',
@@ -266,8 +271,8 @@ export default {
                 hslH : 0,
                 hslS : 0,
                 hslL : 0,
-                hsvS : -1,
-                hsvV : 0,
+                // hsvS : -1,
+                // hsvV : 0,
                 picking : false,
                 panel : {
                     w : 0,
@@ -345,6 +350,8 @@ export default {
         },
         _hexChange : function (value) {
 
+            value = value.trim();
+
             if (value.length !== 7 &&
                 value.length !== 9) {
 
@@ -386,12 +393,24 @@ export default {
             } catch (e) {}
 
         },
-        _rgbaChangeR : function (value) {
+        _rgbaChangeR : function () {
+
+            let red = +this.data.colorValue.r || 0;
+
+            if (red < 0) {
+
+                red = 0;
+
+            } else if (red > maxHex) {
+
+                red = maxHex;
+
+            }
 
             try {
 
                 let hsl = this.colorObj
-                    .red(value)
+                    .red(red)
                     .hsl();
 
                 if (this.data.hslHSync) {
@@ -406,12 +425,24 @@ export default {
             } catch (e) {}
 
         },
-        _rgbaChangeG : function (value) {
+        _rgbaChangeG : function () {
+
+            let green = +this.data.colorValue.g || 0;
+
+            if (green < 0) {
+
+                green = 0;
+
+            } else if (green > maxHex) {
+
+                green = maxHex;
+
+            }
 
             try {
 
                 let hsl = this.colorObj
-                    .green(value)
+                    .green(green)
                     .hsl();
 
                 if (this.data.hslHSync) {
@@ -426,12 +457,24 @@ export default {
             } catch (e) {}
 
         },
-        _rgbaChangeB : function (value) {
+        _rgbaChangeB : function () {
+
+            let blue = +this.data.colorValue.b || 0;
+
+            if (blue < 0) {
+
+                blue = 0;
+
+            } else if (blue > maxHex) {
+
+                blue = maxHex;
+
+            }
 
             try {
 
                 let hsl = this.colorObj
-                    .blue(value)
+                    .blue(blue)
                     .hsl();
 
                 if (this.data.hslHSync) {
@@ -446,11 +489,23 @@ export default {
             } catch (e) {}
 
         },
-        _hslChangeH : function (value) {
+        _hslChangeH : function () {
 
             if (!this.data.hslHSync) {
 
                 return;
+
+            }
+
+            let value = +this.data.colorValue.h || 0;
+
+            if (value < 0) {
+
+                value = 0;
+
+            } else if (value > num360) {
+
+                value = num360 - 1;
 
             }
 
@@ -461,7 +516,7 @@ export default {
                     .object();
 
                 hslObj.h = value;
-               
+                
                 this.data.hslH = color(hslObj)
                     .hsl()
                     .object()
@@ -470,7 +525,19 @@ export default {
             } catch (e) {}
 
         },
-        _hslChangeS : function (value) {
+        _hslChangeS : function () {
+
+            let value = +(this.data.colorValue.s.replace('%', '')) || 0;
+
+            if (value < 0) {
+
+                value = 0;
+
+            } else if (value > num100) {
+
+                value = num100;
+
+            }
 
             try {
 
@@ -478,7 +545,7 @@ export default {
                     .hsl()
                     .object();
 
-                hslObj.s = +value.replace('%', '');
+                hslObj.s = value;
                
                 this.data.hslS = color(hslObj)
                     .hsl()
@@ -488,7 +555,19 @@ export default {
             } catch (e) {}
 
         },
-        _hslChangeL : function (value) {
+        _hslChangeL : function () {
+
+            let value = +(this.data.colorValue.l.replace('%', '')) || 0;
+
+            if (value < 0) {
+
+                value = 0;
+
+            } else if (value > num100) {
+
+                value = num100;
+
+            }
 
             try {
 
@@ -496,7 +575,7 @@ export default {
                     .hsl()
                     .object();
 
-                hslObj.l = +value.replace('%', '');
+                hslObj.l = value;
                
                 this.data.hslL = color(hslObj)
                     .hsl()
@@ -506,32 +585,12 @@ export default {
             } catch (e) {}
 
         },
-        _alphaChange : function (value) {
+        _alphaChange : function () {
 
-            if (isNaN(+value) ||
-                !this.conf.allowAlpha) {
-
-                return;
-
-            }
-
-            if (value > maxAlpha) {
-
-                value = maxAlpha;
-
-            }
-
-            if (value < 0) {
-
-                value = 0;
-
-            }
-
-            this.data.alpha = value;
             this.$emit('alpha-slider-change');
 
         },
-        _alphaChangePer : function (per) {
+        _alphaChangePer : function () {
 
             if (!this.conf.allowAlpha) {
 
@@ -539,15 +598,15 @@ export default {
 
             }
 
-            if (per > 1) {
-
-                per = 1;
-
-            }
+            let per = +this.data.colorValue.a || 0;
 
             if (per < 0) {
 
                 per = 0;
+
+            } else if (per > 1) {
+
+                per = 1;
 
             }
 
@@ -568,13 +627,13 @@ export default {
         },
         _hsvChangeSV : function (s, v) {
 
-            this.data.hsvS = s;
-            this.data.hsvV = v;
+            // this.data.hsvS = s;
+            // this.data.hsvV = v;
 
             let hsl = color({
                 h : this.data.hslH,
                 s : s,
-                v : num100 - v
+                v : v
             })
                 .hsl()
                 .object();
@@ -587,6 +646,12 @@ export default {
           
             this.data.dontPickColor = true;
             this._moveItemRecord(0);
+
+        },
+        _stopmoveStraw : function () {
+
+            this.data.dontPickColor = false;
+            this._moveMouseup();
 
         },
         _rePositionStrawWithSV : function (s, v) {
@@ -620,7 +685,7 @@ export default {
             };
             this._hsvChangeSV(
                 (((+x) + (this.data.strawSize / 2)) / this.data.panel.w) * num100,
-                (((+y) + (this.data.strawSize / 2)) / this.data.panel.h) * num100
+                (num100 - ((((+y) + (this.data.strawSize / 2)) / this.data.panel.h) * num100))
             );
 
             this.Vue.nextTick(() => {
@@ -675,7 +740,7 @@ export default {
                     r : Math.round(colorObj.red()),
                     g : Math.round(colorObj.green()),
                     b : Math.round(colorObj.blue()),
-                    a : this.alphaPer || 1
+                    a : isNaN(this.alphaPer) ? 1 : this.alphaPer
                 };
 
             } else if (type === 'hsla') {
@@ -686,7 +751,7 @@ export default {
                     h : Math.round(hsl.h),
                     s : `${Math.round(hsl.s)}%`,
                     l : `${Math.round(hsl.l)}%`,
-                    a : this.alphaPer || 1
+                    a : isNaN(this.alphaPer) ? 1 : this.alphaPer
                 };
 
             }
@@ -782,6 +847,17 @@ export default {
             }, copyShowtime);
 
         },
+        set : function (value) {
+
+            this._hslHSync(true);
+
+            let result = this._set(value);
+
+            this.Vue.nextTick(() => this._hslHSync(false));
+
+            return result;
+
+        },
         togglePicker : function (show) {
 
             if (this.data.toggling) {
@@ -832,7 +908,9 @@ export default {
         this.Move.target = '.straw';
         this.Move.container = '.panel';
 
+        this._hslHSync(true);
         this._syncColorFromValue();
+        this._hslHSync(false);
 
         this.Vue.nextTick(() => {
 
@@ -850,7 +928,7 @@ export default {
 
         this.$watch('data.hslH', () => {
 
-            this.data.hslHReversal = Math.floor(this.data.hslH);
+            this.data.hslHReversal = this.data.hslH;
 
         }, {
             immediate : true
@@ -880,6 +958,8 @@ export default {
             immediate : true
         });
 
+        window.color = color;
+
         this.$watch('colorHex', () => {
 
             let colorObj = this.colorObj;
@@ -897,15 +977,8 @@ export default {
 
             }
 
-            let s = this.data.hsvS;
-            let v = this.data.hsvV;
-
-            if (s === -1) {
-
-                s = hsv.s;
-                v = num100 - hsv.v;
-
-            }
+            let s = hsv.s;
+            let v = num100 - hsv.v;
 
             this._rePositionStrawWithSV(s, v);
 
@@ -960,7 +1033,7 @@ export default {
             this.data.dontPickColor = false;
             this._hsvChangeSV(
                 (((+this.Move.current.x) + (this.data.strawSize / 2)) / this.data.panel.w) * num100,
-                (((+this.Move.current.y) + (this.data.strawSize / 2)) / this.data.panel.h) * num100
+                (num100 - ((((+this.Move.current.y) + (this.data.strawSize / 2)) / this.data.panel.h) * num100))
             );
 
         });
@@ -969,7 +1042,7 @@ export default {
 
             this._hsvChangeSV(
                 (((+this.Move.current.x) + (this.data.strawSize / 2)) / this.data.panel.w) * num100,
-                (((+this.Move.current.y) + (this.data.strawSize / 2)) / this.data.panel.h) * num100
+                (num100 - ((((+this.Move.current.y) + (this.data.strawSize / 2)) / this.data.panel.h) * num100))
             );
 
         });