+++ src/lib/components/index.js
@@ -51,6 +51,7 @@ import pagination                   from './pagination/index.vue';
 import dialog                       from './dialog/index.vue';
 import calendar                     from './calendar/index.vue';
 import load                         from './load/index.vue';
+import progress                     from './progress/index.vue';
 import sticky                       from './sticky/index.vue';
 import breadcrumbs                  from './breadcrumbs/index.vue';
 import tip                          from './tip/index.vue';
@@ -133,6 +134,7 @@ export default {
     dialog,
     calendar,
     load,
+    progress,
     sticky,
     breadcrumbs,
     tip,